Refacto addIdentity flow to remove RenderProps
